+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed error catching in `retry.DoTxWithResult`
+
 ## v3.115.2
 * Fixed bug with wrong literal YQL representation for signed integer YDB types